LayoutTests/imported/w3c/web-platform-tests/streams/readable-byte-streams/properties.serviceworker.https-expected.txt


PASS Service worker test setup 
PASS Can get the ReadableStreamBYOBReader constructor indirectly 
PASS ReadableStreamBYOBReader constructor should get a ReadableStream object as argument 
PASS ReadableStreamBYOBReader instances should have the correct list of properties 
FAIL ReadableStreamBYOBRequest instances should have the correct list of properties assert_equals: constructor has 0 parameters expected 0 but got 2
FAIL ReadableByteStreamController instances should have the correct list of properties assert_equals: constructor has no parameters expected 0 but got 3
